Rearrange settings classes in correct files
